requirements.txt
PySide6==6.6.2
reportlab==4.1.0
Pillow==10.2.0
